+++ src/io/index.ts
@@ -1,2 +1,3 @@
 export * from './fileProvider';
+export * from './storageProvider';
 export * from './userInteractionProvider';