+++ src/utils/index.ts
@@ -2,3 +2,4 @@ export * from './disposeProvider';
 export * from './dateUtils';
 export * from './httpRegionUtils';
 export * from './notebookUtils';
+export * from './textEditorUtils';